Add migrations for currencies table

+++ migrations/2_account_table.up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE accounts (
        id uuid DEFAULT uuid_generate_v4() NOT NULL,
-       currency_numeric_code integer NOT NULL,
+       currency_numeric_code smallint NOT NULL,
        balance numeric NOT NULL
 );
